OpenAIComboMoDrAg/code/requirements.txt
bitsandbytes
pubchempy
rdkit
chembl_webresource_client
rcsb-api
deepchem
dockstring
openbabel-wheel
openai
langchain_core 
langchain_openai
langgraph
gradio
torch
matplotlib
pillow 
gradio-client
transformers
dockstring
openbabel-wheel
numpy
elevenlabs
lightgbm
tf-keras
tensorflow
accelerate